singly lists

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -10,7 +10,7 @@
 
 size_t print_listint(const listint_t *h)
 {
-	int n = 0;
+	size_t n = 0;
 
 	while (h != NULL)
 	{
